// src/api/stetsom/model/bannerStatus.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */

export type BannerStatus = (typeof BannerStatus)[keyof typeof BannerStatus];

export const BannerStatus = {
  ACTIVE: "ACTIVE",
  INACTIVE: "INACTIVE",
  SCHEDULED: "SCHEDULED",
} as const;